LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/access-control-preflight-request-invalid-status-400-expected.txt

CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/XMLHttpRequest/resources/access-control-preflight-request-invalid-status.py?code=400 due to access control checks.

PASS Request with 400 preflight response 
